refactor(mysql): organize connection lifecycle tests
